// apps/survivor/src/core/HomesteadModels.generated.ts
// Generated by scripts/lib/survivor-homestead.mjs from the licensed source meshes.
export const HOMESTEAD_MODELS = {
    "Inn": {
        "width": 8,
        "depth": 7.984609562807982,
        "height": 6.931949231885001
    },
    "Blacksmith": {
        "width": 7,
        "depth": 5.908568199653493,
        "height": 5.40456591463648
    },
    "House_1": {
        "width": 4.838054095506116,
        "depth": 6,
        "height": 7.64999971490719
    },
    "Well": {
        "width": 1.3280116016775203,
        "depth": 2,
        "height": 2.492034403889249
    }
} as const;